feat: refactor import paths for constants and initialize game state structure
