+++ app/core/constants/init-game-state.ts
